+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## V0.x-dev
 - Make configuration directory a constant
 - Change configuration directory from ~/.wp-sync-util to ~/.wp-site-sync
+- Add return code checking for ssh, scp and rsync commands
 
 ## V0.1.1
 - Added README.md file